LayoutTests/platform/wpe/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/datetime-local-trailing-zeros-expected.txt



FAIL Verifies that trailing zeros in the milliseconds portion of the date strings are removed. assert_equals: expected "2022-04-19T12:34:56.01" but got "2022-04-19T12:34:56.010"
